refactor: refactor tooltip
