+++ packages/popover/vue-popper.js
@@ -131,7 +131,6 @@ function useVuePopper(props, { emit, slots, referenceEl }) {
   }
 
   function updatePopper() {
-    debugger
     const popperJSRef = popperJS.value
     if (popperJSRef) {
       popperJSRef.update()
