+++ packages/tooltip/index.js
@@ -1,5 +1,4 @@
-// import Tooltip from './src/main.vue'
-import Tooltip from './src/main-test'
+import Tooltip from './src/main'
 
 /* istanbul ignore next */
 Tooltip.install = function (app) {
